*: Unify policy and attestation searchers
It's preferable to unify the policy and attestation searcher flows into
a single interface that can be implemented together as the forthcoming
cache searcher needs to share state between the two refs.
