+++ internal/policy/searcher.go
@@ -7,23 +7,29 @@ import (
 	"fmt"
 	"log/slog"
 
+	"github.com/gittuf/gittuf/internal/attestations"
 	"github.com/gittuf/gittuf/internal/gitinterface"
 	"github.com/gittuf/gittuf/internal/rsl"
 )
 
-// Searcher defines the interface for identifying the applicable policy entry in
-// the RSL for some entry.
-type Searcher interface {
+// searcher defines the interface for identifying the applicable policy and
+// attestation entries in the RSL for some entry.
+type searcher interface {
 	FindPolicyEntryFor(rsl.Entry) (*rsl.ReferenceEntry, error)
+	FindAttestationsEntryFor(rsl.Entry) (*rsl.ReferenceEntry, error)
 }
 
-// RegularSearcher implements the Searcher interface. It walks back the RSL from
-// the specified entry to find the latest policy entry.
-type RegularSearcher struct {
+func newSearcher(repo *gitinterface.Repository) searcher {
+	return newRegularSearcher(repo)
+}
+
+// regularSearcher implements the searcher interface. It walks back the RSL from
+// the specified entry to find the latest policy or attestations entry.
+type regularSearcher struct {
 	repo *gitinterface.Repository
 }
 
-func (r *RegularSearcher) FindPolicyEntryFor(entry rsl.Entry) (*rsl.ReferenceEntry, error) {
+func (r *regularSearcher) FindPolicyEntryFor(entry rsl.Entry) (*rsl.ReferenceEntry, error) {
 	// If the requested entry itself is for the policy ref, return as is
 	if entry, isReferenceEntry := entry.(*rsl.ReferenceEntry); isReferenceEntry && entry.RefName == PolicyRef {
 		slog.Debug(fmt.Sprintf("Initial entry '%s' is for gittuf policy, setting that as current policy...", entry.GetID().String()))
@@ -44,6 +50,28 @@ func (r *RegularSearcher) FindPolicyEntryFor(entry rsl.Entry) (*rsl.ReferenceEnt
 	return policyEntry, nil
 }
 
-func NewRegularSearcher(repo *gitinterface.Repository) *RegularSearcher {
-	return &RegularSearcher{repo: repo}
+func (r *regularSearcher) FindAttestationsEntryFor(entry rsl.Entry) (*rsl.ReferenceEntry, error) {
+	// If the requested entry itself is for the attestations ref, return as is
+	if entry, isReferenceEntry := entry.(*rsl.ReferenceEntry); isReferenceEntry && entry.RefName == attestations.Ref {
+		slog.Debug(fmt.Sprintf("Initial entry '%s' is for attestations, setting that as current set of attestations...", entry.GetID().String()))
+		return entry, nil
+	}
+
+	attestationsEntry, _, err := rsl.GetLatestReferenceEntry(r.repo, rsl.ForReference(attestations.Ref), rsl.BeforeEntryID(entry.GetID()))
+	if err != nil {
+		if errors.Is(err, rsl.ErrRSLEntryNotFound) {
+			// Attestations may not be used yet, they're not
+			// compulsory
+			slog.Debug(fmt.Sprintf("No attestations found before initial entry '%s'", entry.GetID().String()))
+			return nil, attestations.ErrAttestationsNotFound
+		}
+
+		return nil, err
+	}
+
+	return attestationsEntry, nil
+}
+
+func newRegularSearcher(repo *gitinterface.Repository) *regularSearcher {
+	return &regularSearcher{repo: repo}
 }
