+++ internal/policy/searcher_test.go
@@ -5,6 +5,7 @@ package policy
 import (
 	"testing"
 
+	"github.com/gittuf/gittuf/internal/attestations"
 	"github.com/gittuf/gittuf/internal/gitinterface"
 	"github.com/gittuf/gittuf/internal/rsl"
 	"github.com/stretchr/testify/assert"
@@ -28,7 +29,7 @@ func TestRegularSearcher(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		searcher := NewRegularSearcher(repo)
+		searcher := newRegularSearcher(repo)
 		policyEntry, err := searcher.FindPolicyEntryFor(entry)
 		assert.Nil(t, err)
 		assert.Equal(t, expectedPolicyEntry.GetID(), policyEntry.GetID())
@@ -71,9 +72,82 @@ func TestRegularSearcher(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		searcher := NewRegularSearcher(repo)
+		searcher := newRegularSearcher(repo)
 		policyEntry, err := searcher.FindPolicyEntryFor(entry)
 		assert.ErrorIs(t, err, ErrPolicyNotFound)
 		assert.Nil(t, policyEntry)
+	})
+
+	t.Run("attestations exist", func(t *testing.T) {
+		tmpDir := t.TempDir()
+		repo := gitinterface.CreateTestGitRepository(t, tmpDir, false)
+
+		currentAttestations := &attestations.Attestations{}
+		if err := currentAttestations.Commit(repo, "Initial attestations\n", false); err != nil {
+			t.Fatal(err)
+		}
+
+		expectedAttestationsEntry, err := rsl.GetLatestEntry(repo)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// Add an entry after
+		if err := rsl.NewReferenceEntry("refs/heads/main", gitinterface.ZeroHash).Commit(repo, false); err != nil {
+			t.Fatal(err)
+		}
+
+		entry, err := rsl.GetLatestEntry(repo)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		searcher := newRegularSearcher(repo)
+		attestationsEntry, err := searcher.FindAttestationsEntryFor(entry)
+		assert.Nil(t, err)
+		assert.Equal(t, expectedAttestationsEntry.GetID(), attestationsEntry.GetID())
+
+		// Try with annotation
+		if err := rsl.NewAnnotationEntry([]gitinterface.Hash{entry.GetID()}, false, "Annotation\n").Commit(repo, false); err != nil {
+			t.Fatal(err)
+		}
+
+		attestationsEntry, err = searcher.FindAttestationsEntryFor(entry)
+		assert.Nil(t, err)
+		assert.Equal(t, expectedAttestationsEntry.GetID(), attestationsEntry.GetID())
+
+		// Requested entry is annotations entry
+		if err := rsl.NewReferenceEntry(attestations.Ref, gitinterface.ZeroHash).Commit(repo, false); err != nil {
+			t.Fatal(err)
+		}
+
+		entry, err = rsl.GetLatestEntry(repo)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		attestationsEntry, err = searcher.FindAttestationsEntryFor(entry)
+		assert.Nil(t, err)
+		assert.Equal(t, entry.GetID(), attestationsEntry.GetID())
+	})
+
+	t.Run("attestations do not exist", func(t *testing.T) {
+		tmpDir := t.TempDir()
+		repo := gitinterface.CreateTestGitRepository(t, tmpDir, false)
+
+		// Add an entry after
+		if err := rsl.NewReferenceEntry("refs/heads/main", gitinterface.ZeroHash).Commit(repo, false); err != nil {
+			t.Fatal(err)
+		}
+
+		entry, err := rsl.GetLatestEntry(repo)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		searcher := newRegularSearcher(repo)
+		attestationsEntry, err := searcher.FindAttestationsEntryFor(entry)
+		assert.ErrorIs(t, err, attestations.ErrAttestationsNotFound)
+		assert.Nil(t, attestationsEntry)
 	})
 }
